+++ Package.swift
@@ -5,7 +5,11 @@ import PackageDescription
 let package = Package(
     name: "ChessKit",
     platforms: [
-        .iOS(.v16), .watchOS(.v9), .macOS(.v13), .tvOS(.v16)
+        .iOS(.v13),
+        .macCatalyst(.v13),
+        .macOS(.v10_15),
+        .tvOS(.v13),
+        .watchOS(.v6)
     ],
     products: [
         .library(