+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # [unreleased]
 
 #### Improvements
-* Enabled `chesskit-swift` to run on oldest platform possible without code changes.
+* Enable `chesskit-swift` to run on oldest platform possible without code changes.
   * Now works on iOS 13+, macOS 10.15+, tvOS 13+, watchOS 6+
+
+#### Bug Fixes
+* Fix `MoveTree` not properly publishing changes via `Game`.
+* Fix `Board.EndResult.repetition` spelling.
+  * This isn't made available yet but will be implemented in an upcoming release.
+
 
 # ChessKit 0.5.0
 Released Sunday, April 14, 2024.